Completed
This is a simple Task Manager.

It stores passwords hashed in text files. We will move everything over into a database in a later project.

This project does allow users to search and sort tasks by type, name, and description.

This project does allow users to create an account. It will NOT send an email to the users after their account is created. An additional library will be used in another project to accomplish this.